requirements.txt
django==1.5.2
django-postmark==0.1.6
django-userena==1.2.3
django-social-auth==0.7.25
django-bootstrap-form==3.0.0
django-devserver==0.6.2
django-compressor==1.3
djangorestframework==2.3.7
django-filter==0.7
python-dateutil==1.5
pillow==2.1.0
south==0.8.2
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==17.5
raven==3.4.1
markdown==2.3.1
